# openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /flightplan_approval_check/{flightplanId}:
    get:
      operationId: get_flightplan_approval_check_flightplanId_flightplan_approval_check__flightplanId__get
      parameters:
      - in: path
        name: flightplanId
        required: true
        schema:
          title: Flightplanid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Flightplan Approval Check Flightplanid
      tags:
      - Extra Routes
      x-description: Verifies possible collisions between a new flightplan before
        approval
...